project_src/requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.0
Flask-Mail==0.9.0
Flask-OpenID==1.2.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
Whoosh==2.6.0
blinker==1.3
coverage==3.7.1
decorator==3.4.0
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
pbr==0.10.0
python-openid==2.2.5
pytz==2014.7
six==1.8.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlparse==0.1.11
Flask-Admin==1.4.2
Flask-Security==1.7.5
yunpian-sdk-python
-e git://github.com/GeeTeam/gt-python-sdk.git#egg=geetest
pandas==0.19.2
xlrd==1.0.0
tornado==4.4.2
psycopg2==2.7
flask-migrate==2.0.3
flask_script==2.0.5
kafka-python